requirements.txt
#Crawler
selenium
websocket-client

# data handling
numpy
pandas>=1.1.5
stockstats
yfinance
#pyfolio
sklearn
#torch-sparse
#torch-geometric
ta

# plot
matplotlib
plotly

# Model Building Requirements
scikit-learn>=0.21.0
gym>=0.17
stable-baselines3[extra]

# testing requirements
pytest
notebook

# packaging
setuptools>=41.4.0
wheel>=0.33.6

keras
mplfinance
finrl

#Web Crawler
requests==2.22.0 
beautifulsoup4==4.8.1

h5py
tensorflow-gpu==2.6.1
xgboost
matplotlib
tqdm
torch
